AVRO-2079: Incorporate remarks from code review
- Reorder imports in TestJava8TimeConversions.java
- Directly cast epochDays to int
- Directly cast millis to int
- Explicitly format java8 and joda time dates in test
- Rename DateTimeLogicalTypeType to DateTimeLogicalTypeImplementation
- Refactor registration of conversions
- Make DateTimeLigicalTypeImplementation field final
- Use parent POM version in Maven plugin tests
- Log warning message about unknown parameter values
- Remove unused imports

+++ lang/java/avro/src/main/java/org/apache/avro/data/Java8TimeConversions.java
@@ -49,7 +49,7 @@ public class Java8TimeConversions {
     public Integer toInt(LocalDate date, Schema schema, LogicalType type) {
       long epochDays = date.toEpochDay();
 
-      return Math.toIntExact(epochDays);
+      return (int) epochDays;
     }
   }
 
@@ -71,7 +71,7 @@ public class Java8TimeConversions {
 
     @Override
     public Integer toInt(LocalTime time, Schema schema, LogicalType type) {
-      return Math.toIntExact(TimeUnit.NANOSECONDS.toMillis(time.toNanoOfDay()));
+      return (int) TimeUnit.NANOSECONDS.toMillis(time.toNanoOfDay());
     }
   }
